requirements.txt
bcrypt==3.1.7
cffi==1.13.1
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
libmagic==1.0
MarkupSafe==1.1.1
Pillow==6.2.1
pycparser==2.19
PyJWT==1.4.2
python-magic==0.4.15
python-magic-bin==0.4.14
six==1.12.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
WTForms==2.2.1
